# flag-backend-master/src/main/resources/META-INF/openapi.yaml
openapi: 3.0.0
info:
  title: DCTx FLAG API
  description: >-
    API for FLAG Web Services
  version: "1.0.0"
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
security:
  - BearerAuth: []
paths:
  /auth/login:
    post:
      tags:
        - public
      summary: >-
        Authentication endpoint. This requires a username and password. This
        will return an authentication token required for accessing the other
        endpoints.
      operationId: auth
      requestBody:
        description: >-
          Login request body contains the username and password
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        200:
          description: >-
            A successful login will return an access code which is a bearer token
            needed to access the other endpoints.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        default:
          description: >-
            An error occured while performing the operation. A details JSON error is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/client/token:
    post:
      tags:
        - public
      summary: >-
        Authentication endpoint. This requires a client ID and secret. This
        will return an access token required for accessing the other
        endpoints.
      operationId: authClientToken
      requestBody:
        description: >-
          Login request body contains the client ID and client secret
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientTokenRequest'
      responses:
        200:
          description: >-
            A successful login will return an access token which is a bearer token
            needed to access the other endpoints.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientTokenResponse'
        default:
          description: >-
            An error occured while performing the operation. A details JSON error is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /donations:
    get:
      tags:
        - donation
      summary: >-
        Retrieve all donations
      operationId: getDonations
      parameters:
      responses:
        200:
          description: >-
            The details of the donations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Donation'
        default:
          description: >-
            An error occured while performing the operation. A details JSON error is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - donation
      summary: Creates a donation.
      operationId: createDonation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Donation'
      responses:
        200:
          description: >-
            The details of the donation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Donation'
        default:
          description: >-
            An error occured while performing the operation. A details JSON error is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /donations/{id}:
    get:
      tags:
        - donation
      summary: >-
        Retrieve donation by Id
      operationId: getDonationById
      parameters:
          - name: id
            in: path
            required: true
            type: number
            description: Id of the donation
            example: 1
      responses:
        200:
          description: >-
            The details of the donation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Donation'
        default:
          description: >-
            An error occured while performing the operation. A details JSON error is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

                
    delete:
      tags:
        - donation
      summary: >-
        Delete donation by Id
      operationId: getDonationById
      parameters:
          - name: id
            in: path
            required: true
            type: number
            description: Id of the donation
            example: 0
      responses:
        200:
          description: >-
            Donation deleted
        default:
          description: >-
            An error occured while performing the operation. A details JSON error is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /donations/request/{id}:
    get:
      tags:
        - donation
      summary: >-
        Retrieve donation by request Id
      operationId: getDonationByRequestId
      parameters:
          - name: id
            in: path
            required: true
            type: number
            description: Id of the donation request
            example: 1
      responses:
        200:
          description: >-
            The details of the donation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Donation'
        default:
          description: >-
            An error occured while performing the operation. A details JSON error is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    Error:
      type: object
      properties:
        id:
          type: string
          description: >-
            a unique identifier for this particular occurrence of the problem.
        links:
          type: object
          description: >-
            a links object containing the following members;
          properties:
            about:
              type: string
              description: >-
                a link that leads to further details about this particular occurrence of the problem.
          additionalProperties: true
        status:
          type: string
          description: >-
            the HTTP status code applicable to this problem, expressed as a
            string value.
        code:
          type: string
          description: >-
            an application-specific error code, expressed as a string value.
        title:
          type: string
          description: >-
            a short, human-readable summary of the problem that SHOULD NOT
            change from occurrence to occurrence of the problem, except for
            purposes of localization.
        detail:
          type: string
          description: >-
            a human-readable explanation specific to this occurrence of the
            problem. Like title, this field’s value can be localized.
        source:
          type: array
          description: >-
            an object containing references to the source of the error, optionally including any of the following members;
          items:
            type: object
            properties:
              pointer:
                type: string
                description: >-
                  a JSON Pointer [RFC6901] to the associated entity in the
                  request document [e.g. "/data" for a primary data object, or "
                  /data/attributes/title" for a specific attribute].
              parameter:
                type: string
                description: >-
                  a string indicating which URI query parameter caused the error
                  .
        meta:
          type: object
          description: >-
            a meta object containing non-standard meta-information about the error.
          additionalProperties: true
    LoginRequest:
      type: object
      properties:
        username:
          type: string
          description: >-
            The username to login.
          example: 'test_user'
        password:
          type: string
          format: password
          description: >-
            The password needed to login.
          example: '123456'
    LoginResponse:
      type: object
      properties:
        access-code:
          type: string
          description: >-
            A bearer token needed to access the protected endpoints
        refresh-code:
          type: string
          description: >-
            A token required to refresh an access token
    ClientTokenRequest:
      type: object
      properties:
        client_id:
          type: string
          description: >-
            The client ID for the apps using the API. This should be requested from the Keycloak admin.
          example: 'frontend-dev'
        client_secret:
          type: string
          description: >-
            The secret for the apps using the API. This should be requested from the Keycloak admin.
          example: '776aabe1-a635-4b36-b635-06b28d778268'
    ClientTokenResponse:
      type: object
      properties:
        access_token:
          type: string
          description: >-
            A bearer token needed to access the protected endpoints
          example: 'eyJhbGciOiJSUzI1NiIsInR5cCIgOiAiSldUIiwia2lk...'
...